+++ backend/fortran/src/z_indexx.f90
@@ -11,7 +11,7 @@ SUBROUTINE z_indexx(n,arr,indx)
    PARAMETER (M=7,NSTACK=50)
    integer(8) i,indxt,ir,itemp,j,jstack,k,l,istack(NSTACK)
    double precision a
-   integer :: allocate_status=0
+   integer(8) :: allocate_status=0
    double precision, dimension(:), allocatable :: arr_0
 !
 !
