+++ josab_tutorial/sim-josab-02.py
@@ -35,7 +35,7 @@ inc_a_x = 450 # Waveguide widths.
 inc_a_y = 200
 # Unit cell dimensions must be sufficiently large to ensure fields are zero at outermost boundary.
 domain_x = 3.01*wl_nm
-domain_y = domain_x #be careful to ensure not whole integer multiples
+domain_y = domain_x #be careful to ensure not whole integer(8) multiples
 inc_shape = 'rectangular' # Shape of the waveguide.
 
 # Specify number of electromagnetic modes and acoustic modes involved in the
